dri3: Don't enable the DRI3 extension unless some screen supports it
There's no reason to advertise this extension unless one of the
hardware drivers actually supports it. Not listing it means it's
slightly easier for users to tell what's going on.

On the other hand, not listing it means we may have applications that
only check for the extension and not for appropriate per-screen
support. I don't think that's a real risk as DRI3 is only useful for
systems with deep knowledge of the hardware.

Signed-off-by: Keith Packard <keithp@keithp.com>
Reviewed-by: Eric Anholt <eric@anholt.net>