miext/shadow: Remove ancient backwards-compatibility hack
Here's a trip down memory lane.  Back when we merged kdrive we adopted
kdrive's version of shadow, which used damage directly instead of
hand-rolling it.  However a couple of Xorg drivers referred to the
accumulated damage region in the shadow private directly, so I added a
hack to copy the damage region around.

That was 9148d8700b7c5afc2644e5820c57c509378f93ce, back in early 2006.
Eight years is unusually patient for me.  The neomagic and trident drivers
were still relying on this, but they've been modified to ask the damage
code for the region instead.

Reviewed-by: Keith Packard <keithp@keithp.com>
Signed-off-by: Adam Jackson <ajax@redhat.com>